requirements.txt
pyautogui
pillow
requests
pywin32
selenium
tqdm
